requirements.txt
requests
pytest
voluptuous
pytest-voluptuous
selene==2.0.0b17
allure-pytest
python-dotenv
curlify